move reparsing tests
